fixed b64decode call to take new param.
