+++ Testing/k6/inc/common.js
@@ -17,7 +17,7 @@ import http from "k6/http";
 import { b64decode } from "k6/encoding";
 import { check, group, sleep } from "k6";
 import { Rate, Trend } from "k6/metrics";
-import * as TestOptions from "./testOptions";
+
 
 export let passwd = __ENV.HG_PASSWORD;
 
@@ -55,7 +55,61 @@ export let NoteUrl = baseUrl + "/v1/api/Note";
 export let UserProfileUrl = baseUrl + "/v1/api/UserProfile";
 
 export let ClientId = __ENV.HG_CLIENT ? __ENV.HG_CLIENT : "k6"; // default to k6 client id
-
+//-------------------------------------------------------------------------
+export let loadOptions = {
+    vu: maxVus,
+    stages: [
+        { duration: "2m", target: rampVus }, // simulate ramp-up of traffic from 1 users over a few minutes.
+        { duration: "3m", target: rampVus }, // stay at number of users for several minutes
+        { duration: "3m", target: maxVus }, // ramp-up to users peak for some minutes (peak hour starts)
+        { duration: "3m", target: maxVus }, // stay at users for short amount of time (peak hour)
+        { duration: "2m", target: rampVus }, // ramp-down to lower users over 3 minutes (peak hour ends)
+        { duration: "3m", target: rampVus }, // continue for additional time
+        { duration: "2m", target: 0 }, // ramp-down to 0 users
+    ],
+    thresholds: {
+        errors: ["rate < 0.05"], // threshold on a custom metric
+        http_req_duration: ["p(90)< 9000"], // 90% of requests must complete this threshold
+        http_req_duration: ["avg < 5000"], // average of requests must complete within this time
+    },
+};
+export let smokeOptions = {
+    vus: 1,
+    iterations: 1,
+};
+export let soakOptions = {
+    stages: [
+        { duration: "1m", target: 10 }, // below normal load
+        { duration: "2m", target: 250 },
+        { duration: "3h56m", target: 250 }, // stay at high users for hours 'soaking' the system
+        { duration: "2m", target: 0 }, // drop back down
+    ],
+};
+export let spikeOptions = {
+    stages: [
+        { duration: "20s", target: 10 }, // below normal load
+        { duration: "1m", target: 10 },
+        { duration: "1m", target: 400 }, // spike to super high users
+        { duration: "5m", target: 400 }, // stay there
+        { duration: "1m", target: 200 }, // scale down
+        { duration: "3m", target: 10 },
+        { duration: "10s", target: 0 }, //
+    ],
+};
+export let stressOptions = {
+    stages: [
+        { duration: "2m", target: 50 }, // below normal load
+        { duration: "5m", target: 100 },
+        { duration: "2m", target: 200 }, // normal load
+        { duration: "5m", target: 200 },
+        { duration: "2m", target: 400 }, // around the breaking point
+        { duration: "4m", target: 400 },
+        { duration: "2m", target: 500 }, // beyond the breaking point
+        { duration: "5m", target: 550 },
+        { duration: "5m", target: 0 }, // scale down. Recovery stage.
+    ],
+};
+//----------------------------------------------------------------------
 export let users = [
     {
         username: "loadtest_01",
@@ -123,10 +177,11 @@ export let users = [
     }
 ];
 
-function parseJwt(jwt) {
-    var accessToken = jwt.split(".")[1];
 
-    var decoded = b64decode(accessToken, "rawurl");
+
+function parseJwt(jwt) {
+    var middlepart = String(jwt).split(".")[1];
+    var decoded = b64decode(String(middlepart), "rawurl", 's');
     var token_json = JSON.parse(decoded);
     return token_json;
 }
@@ -134,6 +189,7 @@ function parseJwt(jwt) {
 function parseHdid(accessToken) {
     var json = parseJwt(accessToken);
     var hdid = json["hdid"];
+    console.log("hdid= " + hdid);
     return hdid;
 }
 
@@ -146,21 +202,21 @@ export function groupWithDurationMetric(name, group_function) {
 
 export function OptionConfig() {
     if (OptionsType == "load") {
-        return TestOptions.loadOptions;
+        return loadOptions;
     }
     if (OptionsType === "smoke") {
-        return TestOptions.smokeOptions;
+        return smokeOptions;
     }
     if (OptionsType === "soak") {
-        return TestOptions.soakOptions;
-    }
-    if (OptionsType === "spike") {
-        return TestOptions.soakOptions;
-    }
-    if (OptionsType === "stress") {
         return soakOptions;
     }
-    return TestOptions.smokeOptions;
+    if (OptionsType === "spike") {
+        return spikeOptions;
+    }
+    if (OptionsType === "stress") {
+        return stressOptions;
+    }
+    return smokeOptions;
 }
 
 export function getExpiresTime(seconds) {
