+++ Testing/k6/services/laboratory/docker-compose.yml
@@ -4,8 +4,8 @@ services:
     image: loadimpact/k6:latest
     command: run /services/laboratory/k6_laboratory.js
     volumes:
-      - ../../../src/services/laboratory:/services/laboratory
-      - ../../../src/inc:/inc
+      - ../../services/laboratory:/services/laboratory
+      - ../../inc:/inc
     environment: 
       - "HG_TYPE=${HG_TYPE}"
       - "HG_ENV=${HG_ENV}"
